Reorganize newtmgr commands
* Move commands to cmd package in newtmgr/cmd
* Place each command into its own file
